server/requirements.txt
# Core GUI Framework
PyQt5==5.15.9
PyQt5-tools==5.15.9.3.3

# Voice Recognition & Text-to-Speech
SpeechRecognition==3.10.0
pyttsx3==2.90
PyAudio==0.2.11

# Authentication & Security
PyJWT==2.8.0
cryptography==41.0.7
requests==2.31.0

# Database
pymongo==4.6.0
motor==3.3.2

# AI/ML for Smart Features
google-generativeai==0.8.3
numpy==1.24.3
pandas==2.0.3

# System Integration
psutil==5.9.6
plyer==2.1.0
keyboard==0.13.5

# Development & Utilities
python-dotenv==1.0.0
colorlog==6.7.0
schedule==1.2.0

# Optional: For better system integration
pywin32==306; sys_platform == "win32"